fix: update date formatting in WeatherCard and CurrentCity components for consistency

+++ src/components/WeatherCard/WeatherCard.js
@@ -68,8 +68,8 @@ const WeatherCard = () => {
           feelsLike: data.current.feels_like,
           dateTime: new Date(data.current.dt * 1000).toLocaleString("en-GB", {
             day: "numeric",
-            month: "long",
-            weekday: "long",
+            month: "short",
+            weekday: "short",
             hour: "2-digit",
             minute: "2-digit",
           }),
